Indexing graph correctly, new file created for loading
